requirements.txt
python-telegram-bot==13.15
python-dotenv>=1.0
fpdf2>=2.7.4
qrcode>=7.4.2
Pillow>=10.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
